Trim down commonutils
