+++ src/common/commonutils/CMakeLists.txt
@@ -16,8 +16,7 @@ add_library(commonutils STATIC
     SocketUtils.c
     SshUtils.c
     UrlUtils.c
-    UserUtils.c
-    CommonUtils.cpp)
+    UserUtils.c)
 
 target_compile_options(commonutils PRIVATE -Wno-psabi -Wall -Wextra -Wunused -Werror -Wformat -Wformat-security -Wno-nonnull -Wno-unused-result -Wunused-const-variable=2 -Wunused-macros)
